pyproject.toml
[project]
name = "githubportia"
version = "0.1.0"
description = "LLM-powered resume parser built with Portia AI"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "portia-sdk-python[all]>=0.7.0",
    "pdfminer.six>=20221105",
    "python-dotenv>=1.0.0",
    "PyPDF2",
    "pdfplumber",
    "PyMuPDF",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = []
include = [
    "recruiting_agent/**/*.py",
    "main.py",
    "README.md",
    "env.example",
]

[tool.black]
line-length = 88
target-version = ['py313']

[tool.ruff]
line-length = 88
target-version = "py313"
